Fix 1

+++ include/infinite-color-engine/InfiniteInterpolator.h
@@ -18,8 +18,8 @@ protected:
 public:
 	virtual ~InfiniteInterpolator() = default;
 
-	virtual void setTargetColors(std::vector<linalg::aliases::float3>&& new_rgb_targets, float startTimeMs, bool debug) = 0;
-	virtual void updateCurrentColors(float currentTimeMs, float minBrightness) = 0;
+	virtual void setTargetColors(std::vector<linalg::aliases::float3>&& new_rgb_targets, long long startTimeMs, bool debug) = 0;
+	virtual void updateCurrentColors(long long currentTimeMs, float minBrightness) = 0;
 	virtual SharedOutputColors getCurrentColors(float minBrightness) = 0;
 	virtual void resetState() = 0;
