+++ sources/infinite-color-engine/InfiniteHybridRgbInterpolator.cpp
@@ -70,11 +70,11 @@ void InfiniteHybridRgbInterpolator::resetToColors(std::vector<float3> colors, fl
 	setTargetColors(std::move(colors), startTimeMs);
 }
 
-void InfiniteHybridRgbInterpolator::setTargetColors(std::vector<float3>&& new_rgb_targets, float startTimeMs, bool debug) {
+void InfiniteHybridRgbInterpolator::setTargetColors(std::vector<float3>&& new_rgb_targets, long long startTimeMs, bool debug) {
 	if (new_rgb_targets.empty())
 		return;
 
-	const float delta = (!_isAnimationComplete) ? std::clamp(startTimeMs - _lastUpdate, 0.f, 100.0f) : 0.f;
+	const float delta = (!_isAnimationComplete) ? std::clamp(static_cast<float>(startTimeMs - _lastUpdate), 0.f, 100.0f) : 0.f;
 
 	if (debug)
 	{
@@ -118,20 +118,20 @@ void InfiniteHybridRgbInterpolator::setTargetColors(std::vector<float3>&& new_rg
 
 void InfiniteHybridRgbInterpolator::resetState() {
 	_isAnimationComplete = true;
-	_lastUpdate = 0.0f;
+	_lastUpdate = 0;
 	_targetColorsRGB.clear();
 	_currentColorsRGB.clear();
 	_velocitiesRGB.clear();
 }
 
-void InfiniteHybridRgbInterpolator::updateCurrentColors(float currentTimeMs, float minBrightness) {
+void InfiniteHybridRgbInterpolator::updateCurrentColors(long long currentTimeMs, float minBrightness) {
 	if (_isAnimationComplete)
 	{
 		_lastUpdate = currentTimeMs;
 		return;
 	}
 
-	float dt = std::clamp(currentTimeMs - _lastUpdate, 0.001f, 100.0f);
+	float dt = std::clamp(static_cast<float>(currentTimeMs - _lastUpdate), 0.001f, 100.0f);
 	_lastUpdate = currentTimeMs;
 
 	auto computeChannelVec = [&](float3& cur, const float3& tgt, const float3& diff, float3& vel) -> bool {
